# docs/swagger.yaml
definitions:
...
  project.ScriptGroup:
    properties:
      guid:
        type: string
      objectType:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
...
  project.ScriptRun:
    properties:
      development:
        type: boolean
      doNotRecord:
        type: boolean
      error:
        type: string
      guid:
        type: string
      htmlOutput:
        type: string
      objectType:
        type: string
      percentCompleted:
        type: integer
      requestsMadeCount:
        type: integer
      scriptGroup:
        type: string
      status:
        type: string
      textOutput:
        type: string
      title:
        type: string
      totalRequestCount:
        type: integer
    type: object
...
host: localhost
info:
  contact:
    email: support@forensant.com
    name: API Support
    url: https://forensant.com/support
  description: This provides the common functions which are relied upon by the Proximity
    frontends.
  license:
    name: Commercial
    url: https://proximity.forensant.com/terms
  termsOfService: https://forensant.com/terms/
  title: Proximity Core
  version: "1.0"
paths:
...
  /requests/sitemap:
    get:
      description: gets a list of all paths observed by the proxy
      parameters:
      - description: An optional filter on the query to restrict to specific paths
        in: query
        name: parent
        required: true
        type: string
      - description: An optional filter on the query to restrict to the paths to those
          seen for a particular scan
        in: query
        name: scan_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Gets the sitemap
      tags:
      - Requests
  /script_groups:
    get:
      description: gets a list of all script groups
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/project.ScriptGroup'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get All Script Groups
      tags:
      - Scripting
    post:
      description: adds or updates a script group
      parameters:
      - description: Script Group details in JSON format
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/project.ScriptGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Add/Update Script Group
      tags:
      - Scripting
  /script_groups/{guid}:
    get:
      description: gets a specific script group
      parameters:
      - description: Script group guid
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ScriptGroup'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get Script Group
      tags:
      - Scripting
  /script_groups/{guid}/archive:
    patch:
      description: updates the archived status of a script group
      parameters:
      - description: script group guid
        in: path
        name: guid
        required: true
        type: string
      - description: archive status to set
        in: formData
        name: archive
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Archive Script Group
      tags:
      - Scripting
  /script_groups/{guid}/title:
    patch:
      description: updates the title of a script group
      parameters:
      - description: script group guid
        in: path
        name: guid
        required: true
        type: string
      - description: title to set
        in: formData
        name: title
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Set Script Group Title
      tags:
      - Scripting
  /scripts:
    get:
      description: gets a list of all scripts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/project.ScriptRun'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get All Scripts
      tags:
      - Scripting
  /scripts/{guid}:
    get:
      description: gets a single script
      parameters:
      - description: The GUID of the script to fetch
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get A Script
      tags:
      - Scripting
  /scripts/{guid}/append_html_output:
    post:
      description: appends the given HTML to the HTML output of the script
      parameters:
      - description: The GUID of the script to fetch
        in: path
        name: guid
        required: true
        type: string
      - description: HTML Output to append
        in: body
        name: html
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Append HTML Output for a Script
      tags:
      - Scripting
  /scripts/{guid}/archive:
    patch:
      description: updates the the archived status of a script
      parameters:
      - description: script guid
        in: path
        name: guid
        required: true
        type: string
      - description: archive status to set
        in: formData
        name: archive
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Archive Script
      tags:
      - Scripting
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"